Int to long in repo

+++ src/main/java/com/extia/cinema/repositories/ICategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICategoryRepository extends IGenericRepository<Category, Integer> {
+public interface ICategoryRepository extends IGenericRepository<Category, Long> {
 
     @Query("SELECT c FROM Category c WHERE c.name=?1")
     Category findByName(String name);
